requirements.txt
# ============================================
# Analisador de Microfone com IA
# requirements.txt
# ============================================

# Audio Processing
openai-whisper>=20240930
pyaudio>=0.2.13
numpy>=1.24.0
sounddevice>=0.4.6

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0
python-socketio>=5.9.0
python-engineio>=4.7.0

# AI & NLP
sentence-transformers>=2.2.2
scikit-learn>=1.3.0
thefuzz>=0.19.0
python-Levenshtein>=0.21.0

# Sound Playback
pygame>=2.5.0

# Data Management
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Database
sqlalchemy>=2.0.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
requests>=2.31.0

# Development & Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Logging & Monitoring
# python-logging-loki>=0.3.2  # Optional, pode ter issues

# Performance
psutil>=5.9.0
